gtk3 port

+++ kernel/src/generic/tmpfs.cpp
@@ -16,13 +16,13 @@ bool tmpfs_find_child(tmpfs::tmpfs_node* node, char* name, tmpfs::tmpfs_node** o
     if(node->type != vfs_file_type::directory)
         return false;
 
-    for(std::uint64_t i = 0;i < node->size / sizeof(std::uint64_t);i++ ) {
+    for(std::uint64_t i = 0;i < node->size / sizeof(tmpfs::directory_cont);i++ ) {
 
-        if(node->directory_content[i] == nullptr)
+        if(node->dirents[i].node == nullptr)
             continue;
 
-        if(klibc::strcmp(node->directory_content[i]->name, name) == 0) {
-            *out = node->directory_content[i];
+        if(klibc::strcmp(node->dirents[i].name, name) == 0) {
+            *out = node->dirents[i].node;
             return true;
         }
     }
@@ -139,7 +139,7 @@ std::int32_t tmpfs_readlink(filesystem* fs, char* path, char* buffer) {
         return -EINVAL;
     }
 
-    assert(node->content,"meeeow meeeeeow :3 %s", node->name);
+    assert(node->content,"meeeow meeeeeow :3 %lli", node->ino);
 
     klibc::memcpy(buffer, node->content, 4096);
 
@@ -165,14 +165,23 @@ again:
     }
 
     while(true) {
-        auto current_node = node->directory_content[file->offset / sizeof(tmpfs::tmpfs_node**)];
 
-        file->offset += sizeof(tmpfs::tmpfs_node**);
+        tmpfs::directory_cont* dre = &node->dirents[file->offset / sizeof(tmpfs::directory_cont)];
+
+        auto current_node = dre->node;
+
+        file->offset += sizeof(tmpfs::directory_cont);
+
+        if(file->offset + sizeof(tmpfs::directory_cont) >= node->size) {
+            //klibc::debug_printf("%lli %lli\n", file->offset, node->size);
+            file->vnode.fs->lock.unlock();
+            return current_offset; 
+        }
 
         if(current_node == nullptr)
             goto again;
 
-        if(sizeof(dirent) + klibc::strlen(current_node->name) + 1 > count - current_offset) {
+        if(sizeof(dirent) + klibc::strlen(dre->name) + 1 > count - current_offset) {
             file->vnode.fs->lock.unlock();
             return current_offset; 
         }
@@ -180,16 +189,65 @@ again:
         dirent* current_dir = (dirent*)(out + current_offset);
         current_dir->d_ino = current_node->ino;
         current_dir->d_type = vfs_to_dt_type(current_node->type);
-        current_dir->d_reclen = sizeof(dirent) + klibc::strlen(current_node->name) + 1;
+        current_dir->d_reclen = sizeof(dirent) + klibc::strlen(dre->name) + 1;
         current_dir->d_off = 0;
         current_offset += current_dir->d_reclen;
 
-        klibc::memcpy(current_dir->d_name, current_node->name, klibc::strlen(current_node->name) + 1);
+        klibc::memcpy(current_dir->d_name, dre->name, klibc::strlen(dre->name) + 1);
 
     }
 
     file->vnode.fs->lock.unlock();
     return current_offset;
+}
+
+std::int32_t tmpfs_link(filesystem* fs, char* old_path, char* new_path) {
+    fs->lock.lock();
+
+    tmpfs::tmpfs_node* old_node = tmpfs_lookup(old_path);
+    if(old_node == nullptr) { fs->lock.unlock();
+        return -ENOENT; }
+
+    tmpfs::tmpfs_node* parent = tmpfs_get_parent(new_path);
+    if(parent == nullptr) {
+        fs->lock.unlock();
+        return -ENOENT;
+    }
+
+    if(old_node->type == vfs_file_type::directory) {
+        fs->lock.unlock();
+        return -EPERM;
+    }
+
+    bool is_pasted = false;
+
+    again_paste:
+    for(std::uint64_t i = 0;i < parent->size / sizeof(tmpfs::directory_cont); i++) {
+        if(parent->dirents[i].node == nullptr)  {
+            parent->dirents[i].node = old_node;
+            klibc::memcpy(parent->dirents[i].name, tmpfs_get_name_from_path(new_path), klibc::strlen(tmpfs_get_name_from_path(new_path)) + 1);
+            is_pasted = true;
+            break;
+        }
+    }
+
+    if(is_pasted == false) {
+        alloc_t res = pmm::buddy::alloc(parent->size * sizeof(tmpfs::directory_cont));
+        tmpfs::directory_cont* new_dir = (tmpfs::directory_cont*)(res.phys + etc::hhdm());
+        if(parent->dirents) {
+            klibc::memcpy(new_dir, parent->dirents, parent->size);
+            pmm::buddy::free((std::uint64_t)parent->dirents - etc::hhdm());
+        }
+        parent->dirents = new_dir;
+        parent->physical_size = res.real_size;
+        parent->size = parent->physical_size;
+        goto again_paste;
+    }
+
+    old_node->nlink++;
+    
+    fs->lock.unlock();
+    return 0;
 }
 
 std::uint8_t __tmpfs__create_parent_dirs_by_default = 1; /* Used for ustar */
@@ -223,22 +281,23 @@ std::int32_t tmpfs_create(filesystem* fs, char* path, vfs_file_type type, std::u
     tmpfs::tmpfs_node* new_node = (tmpfs::tmpfs_node*)(pmm::freelist::alloc_4k() + etc::hhdm());
 
 again_paste:
-    for(std::uint64_t i = 0;i < parent->size / sizeof(tmpfs::tmpfs_node**); i++) {
-        if(parent->directory_content[i] == nullptr)  {
-            parent->directory_content[i] = new_node;
+    for(std::uint64_t i = 0;i < parent->size / sizeof(tmpfs::directory_cont); i++) {
+        if(parent->dirents[i].node == nullptr)  {
+            parent->dirents[i].node = new_node;
+            klibc::memcpy(parent->dirents[i].name, tmpfs_get_name_from_path(path), klibc::strlen(tmpfs_get_name_from_path(path)) + 1);
             is_pasted = true;
             break;
         }
     }
 
     if(is_pasted == false) {
-        alloc_t res = pmm::buddy::alloc(parent->size * sizeof(tmpfs::tmpfs_node*));
-        tmpfs::tmpfs_node** new_dir = (tmpfs::tmpfs_node**)(res.phys + etc::hhdm());
-        if(parent->directory_content) {
-            klibc::memcpy(new_dir, parent->directory_content, parent->size);
-            pmm::buddy::free((std::uint64_t)parent->directory_content - etc::hhdm());
+        alloc_t res = pmm::buddy::alloc(parent->size * sizeof(tmpfs::directory_cont));
+        tmpfs::directory_cont* new_dir = (tmpfs::directory_cont*)(res.phys + etc::hhdm());
+        if(parent->dirents) {
+            klibc::memcpy(new_dir, parent->dirents, parent->size);
+            pmm::buddy::free((std::uint64_t)parent->dirents - etc::hhdm());
         }
-        parent->directory_content = new_dir;
+        parent->dirents = new_dir;
         parent->physical_size = res.real_size;
         parent->size = parent->physical_size;
         goto again_paste;
@@ -248,17 +307,80 @@ again_paste:
     new_node->ino = tmpfs_id_ptr++;
     new_node->mode = mode;
     new_node->create_time = time::current_unix_time;
+    new_node->nlink = 1;
 
     if(!fs) {
         if(out)
             *out = new_node;
     }
-    klibc::memcpy(new_node->name, tmpfs_get_name_from_path(path), klibc::strlen(tmpfs_get_name_from_path(path)) + 1);
 
     if(fs)
         fs->lock.unlock();
     return 0;
 }
+
+std::int32_t tmpfs_opt_create(char* path, vfs_file_type type, std::uint32_t mode, char* content, std::uint64_t size, tmpfs::tmpfs_node** out = nullptr) {
+    (void)out;
+    tmpfs::tmpfs_node* parent = tmpfs_get_parent(path);
+
+    if(!parent) {
+        char copy[4096];
+        tmpfs_get_name_parent(path, copy);
+        if(!__tmpfs__create_parent_dirs_by_default) {
+            return -ENOENT;
+        } else {
+            tmpfs_create(nullptr, copy, vfs_file_type::directory, mode, &parent);
+        }
+    }
+
+    if(parent == nullptr) { 
+        return -ENOENT; }
+
+    bool is_pasted = false;
+
+    tmpfs::tmpfs_node* new_node = (tmpfs::tmpfs_node*)(pmm::freelist::alloc_4k() + etc::hhdm());
+
+again_paste:
+    for(std::uint64_t i = 0;i < parent->size / sizeof(tmpfs::directory_cont); i++) {
+        if(parent->dirents[i].node == nullptr)  {
+            parent->dirents[i].node = new_node;
+            klibc::memcpy(parent->dirents[i].name, tmpfs_get_name_from_path(path), klibc::strlen(tmpfs_get_name_from_path(path)) + 1);
+            is_pasted = true;
+            break;
+        }
+    }
+
+    if(is_pasted == false) {
+        alloc_t res = pmm::buddy::alloc(parent->size * sizeof(tmpfs::directory_cont));
+        tmpfs::directory_cont* new_dir = (tmpfs::directory_cont*)(res.phys + etc::hhdm());
+        if(parent->dirents) {
+            klibc::memcpy(new_dir, parent->dirents, parent->size);
+            pmm::buddy::free((std::uint64_t)parent->dirents - etc::hhdm());
+        }
+        parent->dirents = new_dir;
+        parent->physical_size = res.real_size;
+        parent->size = parent->physical_size;
+        goto again_paste;
+    }
+
+    new_node->type = type;
+    new_node->ino = tmpfs_id_ptr++;
+    new_node->mode = mode;
+    new_node->create_time = time::current_unix_time;
+    new_node->nlink = 1;
+
+    if(new_node->type != vfs_file_type::directory) {
+        alloc_t new_content = pmm::buddy::alloc(size);
+        char* new_cont = (char*)(new_content.phys + etc::hhdm());
+        new_node->physical_size = new_content.real_size;
+        new_node->size = size;
+        new_node->content = new_cont;
+        klibc::memcpy(new_cont, content, size);
+    }
+
+    return 0;
+}
+
 
 signed long tmpfs_read(file_descriptor* file, void* buffer, std::size_t count) {
     file->vnode.fs->lock.lock();
@@ -321,40 +443,25 @@ bool tmpfs_test_for_busy(tmpfs::tmpfs_node* node) {
         return true;
 
     if(node->type == vfs_file_type::directory) {
-        for(std::uint64_t i = 0;i < node->size / sizeof(std::uint64_t);i++ ) {
-
-            if(node->directory_content[i] == nullptr)
+        for(std::uint64_t i = 0;i < node->size / sizeof(tmpfs::directory_cont);i++ ) {
+            klibc::debug_printf("grrrrr");
+            assert(node->dirents, "werr %d", node->ino);
+            if(node->dirents[i].node == nullptr)
                 continue;
 
-            if(tmpfs_test_for_busy(node->directory_content[i]) == true)
+            if(tmpfs_test_for_busy(node->dirents[i].node) == true)
                 return true;
         }
     }
     return false;
 }
 
-std::int32_t tmpfs_internal_remove(char* path) {
+std::int32_t tmpfs_internal_remove(tmpfs::tmpfs_node* node) {
 
-    if(klibc::strcmp(path, "/") == 0) {
-        return -EINVAL;
-    }
+    klibc::debug_printf("rm 0x%p", node);
 
-    tmpfs::tmpfs_node* node = tmpfs_lookup(path);
-    if(node == nullptr)
-        return -EINVAL;
-
-    if(tmpfs_test_for_busy(node))
+    if(tmpfs_test_for_busy(node) || node->nlink != 0)
         return -EBUSY;
-
-    tmpfs::tmpfs_node* parent = tmpfs_get_parent(path);
-    if(parent == nullptr)
-        return -EINVAL;
-
-    for(std::uint64_t i = 0;i < parent->size / sizeof(tmpfs::tmpfs_node**); i++) {
-        if(parent->directory_content[i] == node) {
-            parent->directory_content[i] = nullptr;
-        }
-    }
 
     if(node->content) {
         pmm::buddy::free((std::uint64_t)node->content);
@@ -366,10 +473,10 @@ std::int32_t tmpfs_internal_remove(char* path) {
 }
 
 std::int32_t tmpfs_remove(filesystem* fs, char* path) {
-    fs->lock.lock();
-    std::int32_t res = tmpfs_internal_remove(path);
-    fs->lock.unlock();
-    return res;
+    (void)fs;
+    (void)path;
+    assert(0, "h");
+    return 0;
 }
 
 std::int32_t tmpfs_zero(file_descriptor* file) {
@@ -414,10 +521,14 @@ std::int32_t tmpfs_stat(file_descriptor* file, stat* out) {
     out->st_mode = type_to_mode(node->type) | node->mode;
     out->st_size = node->size;
     out->st_ino = node->ino;
+    out->st_nlink = node->nlink;
     out->st_atim.tv_sec = node->access_time;
     out->st_mtim.tv_sec = node->modify_time;
     out->st_ctim.tv_sec = node->create_time;
     node->access_time = time::current_unix_time;
+
+    assert(node->busy_counter >= 0, "Fx");
+
     file->vnode.fs->lock.unlock();
     return 0;
 }
@@ -427,14 +538,25 @@ void tmpfs_close(file_descriptor* file) {
     tmpfs::tmpfs_node* node = (tmpfs::tmpfs_node*)file->fs_specific.tmpfs_pointer;
     node->busy_counter--;
 
-    if(node->busy_counter == 0 && node->should_unlink) { 
-        int res = tmpfs_internal_remove(file->path);
+    assert(node->busy_counter >= 0, "Fb");
+
+    if(node->busy_counter == 0 && node->nlink == 0) { 
+        klibc::debug_printf("rm");
+        int res = tmpfs_internal_remove(node);
         if(res != 0) {
             klibc::debug_printf("rm unlink res %d path %s\n",res, file->path);
         }
     }
 
     file->vnode.fs->lock.unlock();
+}
+
+void tmpfs_ondup(file_descriptor* file) {
+    file->vnode.fs->lock.lock();
+    tmpfs::tmpfs_node* node = (tmpfs::tmpfs_node*)file->fs_specific.tmpfs_pointer;
+    node->busy_counter++;
+    file->vnode.fs->lock.unlock();
+    return;
 }
 
 std::int32_t tmpfs_open(filesystem* fs, void* file_desc, char* path, bool is_directory) {
@@ -449,6 +571,8 @@ std::int32_t tmpfs_open(filesystem* fs, void* file_desc, char* path, bool is_dir
         return -ENOTDIR; }
     
     file_descriptor* fd = (file_descriptor*)file_desc;
+
+    assert(node->busy_counter >= 0, "F");
         
     fd->vnode.fs = fs;
     fd->vnode.stat = tmpfs_stat;
@@ -458,6 +582,7 @@ std::int32_t tmpfs_open(filesystem* fs, void* file_desc, char* path, bool is_dir
     fd->vnode.zero = tmpfs_zero;
     fd->vnode.chmod = tmpfs_chmod;
     fd->vnode.close = tmpfs_close;
+    fd->vnode.ondup = tmpfs_ondup;
     fd->fs_specific.tmpfs_pointer = (std::uint64_t)node;
 
     node->busy_counter++;
@@ -476,12 +601,30 @@ std::int32_t tmpfs_unlink(filesystem* fs, char* path) {
         return -EISDIR;
     }
 
-    node->should_unlink = true;
+    char name[257] = {};
+    klibc::memcpy(name, tmpfs_get_name_from_path(path), klibc::strlen(tmpfs_get_name_from_path(path)) + 1);
 
-    klibc::printf("bus %d\n", node->busy_counter);
+    tmpfs::tmpfs_node* parent = tmpfs_get_parent(path);
 
-    if(node->busy_counter == 0) {
-        int res = tmpfs_internal_remove(path);
+    for(std::uint64_t i = 0;i < parent->size / sizeof(tmpfs::directory_cont); i++) {
+        if(klibc::strcmp(name, parent->dirents[i].name) == 0) {
+            parent->dirents[i].node = nullptr;
+            klibc::memset(parent->dirents[i].name, 0, sizeof(parent->dirents[i].name));
+        }
+    }
+
+    if(parent == nullptr) { 
+        if(fs)
+            fs->lock.unlock();
+        assert(0, "weird things happen...");
+        return -ENOENT; }
+
+    node->nlink--;
+
+    assert(node->busy_counter >= (std::int64_t)0, "v");
+
+    if(node->busy_counter == 0 && node->nlink == 0) {
+        int res = tmpfs_internal_remove(node);
         if(res != 0) {
             klibc::debug_printf("rm unlink res %d path %s\n",res, path);
         }
@@ -498,6 +641,7 @@ void tmpfs::init_default(vfs::node* node) {
     node->fs->create = (int (*)(filesystem *, char *, vfs_file_type, unsigned int))((std::uint64_t)tmpfs_create);
     node->fs->readlink = tmpfs_readlink;
     node->fs->remove = tmpfs_remove;
+    node->fs->link = tmpfs_link;
     node->fs->unlink = tmpfs_unlink;
     klibc::memcpy(node->path, "/\0\0", sizeof("/\0\0") + 1);
 
@@ -506,11 +650,11 @@ void tmpfs::init_default(vfs::node* node) {
     root_node.size = 0;
     root_node.type = vfs_file_type::directory;
     root_node.physical_size = root_alloc.real_size;
+    root_node.nlink = 9999999;
     root_node.size = root_node.physical_size;
-    klibc::memcpy(root_node.name, "/\0", sizeof("/\0") + 1);
     klibc::memcpy(node->internal_path, "/", sizeof("/\0") + 1);
 
-    root_node.directory_content = (tmpfs::tmpfs_node**)(root_alloc.phys + etc::hhdm());
+    root_node.dirents = (tmpfs::directory_cont*)(root_alloc.phys + etc::hhdm());
     log("tmpfs", "root_node is 0x%p",&root_node);
 
 }
